refactor(cockpit): extract determinism gate module
